Add SteadyStateSolverOptions.
For models with pre-equlibrium condition before the main simulation we need to
solve for the steady state du = f(u, p, t) ≈ 0. Added SteadyStateSolverOptions
struct and its constructor getSteadyStateSolverOptions allowing the user to
choose between rootfinding or simulation to find the steady state. The former
uses NonlinearSolve.jl and for the latter simulations can be terminated either
if wrms (weighted root-mean square) or the Newton step is sufficiently small
via a callback to the ODE solver.