Catch gradient error.
For challenging models a subset of chunks for ForwardDiff might fail, but not all chunks. This can result in the retcode being success, while the gradient is incorrect. We not catch this by having a bool variable being set to false if any error occurred when computing the gradient or Hessian.