Add a compute_nllh_and_gradient function
For Bayesian inference several of the frameworks want both the gradient
and the likelihood at the same time. When the gradient has been
computed there is no need to recompute the objective and resolve
the ODE:s. The compute_nllh_and_gradient function only solves
the ODE:s for the gradient.